Encountered a weird equip bug
For some reason, when the module is picked up and I click to equip, the game says it's been removed as if it was already equipped. The result is the stats are lowered to negative values. The weird part is that I can't consistently recreate this bug, so it's hard to find the cause.